chore: resolve React Compiler lint cleanup; flip rules back to error
Fix all 21 sites flagged by react-hooks recommended-latest and remove the
temporary `warn` overrides in eslint.config.js so set-state-in-effect,
preserve-manual-memoization and use-memo revert to `error`.

- use-memo: delete dead hooks/usePerformanceHooks.ts (unreferenced).
- preserve-manual-memoization: align handleUpdateItem useCallback deps to
  `editingItem` in Inventory/List view pages (un-skips page memoization).
- set-state-in-effect: replace reset-in-effect with idiomatic patterns —
  event-handler clear (delete-confirm dialogs), key reset (edit forms, media
  sheets), outer/inner split keyed on load (settings cards, new
  NotificationsCard), and derive-during-render (ExpiryCalendar selection,
  PromoteBar inner split, UserSettingsPage iosHint).
- Genuine external-sync effects (object-URL lifecycle in MediaPreviewSheet /
  useItemImage, timer state machine in useExtractionPoll) keep their effect
  with a targeted, documented eslint-disable-next-line.

Also move the composer edit-mode cancel (x) onto the edit-label row, far right.

Verified: lint (0 errors), tsc, prettier:check, vite build all pass.
