+++ BUGS.md
@@ -5,7 +5,7 @@
 Reported during testing: when the PWA is open and focused, an incoming expiry
 notification does not appear. Needs verification with a **real** data-only FCM
 message before fixing — the foreground `onMessage` handler reads `payload.data.*`,
-so a *notification*-type message (e.g. one sent from the Firebase console test UI)
+so a _notification_-type message (e.g. one sent from the Firebase console test UI)
 is silently ignored, which may fully explain the report. Even for real messages,
 some browsers/OSes suppress a system notification raised while its own tab is
 focused. Likely fix if confirmed: surface foreground messages as an in-app toast
@@ -35,10 +35,3 @@ should carry the target household id and switch the active household to it on
 open (or the inventory route should detect the mismatch and switch) before
 issuing the scoped query. Root cause is the implicit household-context model —
 see "Household context is implicit (LastActiveHouseholdId)…" in `TECH_DEBT.md`.
-
-## Calendar/date picker is not translated or ignores the user's selected language
-
-The calendar (date picker) does not respect the user's selected language — it
-renders untranslated (month/day names, labels) instead of following the active
-i18next locale (`en`/`de`). The picker's locale needs to be wired to the current
-i18n language so it switches alongside the rest of the UI.
